+++ server/app/services/export/CsvExporterService.java
@@ -36,10 +36,8 @@ import services.applicant.ApplicantService;
 import services.applicant.ReadOnlyApplicantProgramService;
 import services.applicant.question.ApplicantQuestion;
 import services.applicant.question.Scalar;
+import services.export.enums.ColumnType;
 import services.pagination.SubmitTimeSequentialAccessPaginationSpec;
-import services.program.Column;
-import services.program.ColumnType;
-import services.program.CsvExportConfig;
 import services.program.ProgramDefinition;
 import services.program.ProgramNotFoundException;
 import services.program.ProgramQuestionDefinition;
